requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
APScheduler
python-dotenv
yagmail
